refactor: all tests in new test folder
